# app/main.py
"""FastAPI application factory and configuration."""

import logging
import os
import secrets
from pathlib import Path
from typing import cast

from fastapi import Depends, FastAPI, Request, status
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import JSONResponse, Response
from fastapi.staticfiles import StaticFiles
from starlette.staticfiles import StaticFiles as StarletteStaticFiles
from starlette.responses import Response as StarletteResponse
from starlette.types import Scope
from slowapi import _rate_limit_exceeded_handler
from slowapi.errors import RateLimitExceeded
from sqlalchemy import exc as sqlalchemy_exc
from sqlalchemy.engine import make_url
from sqlalchemy.ext.asyncio import AsyncSession
from starlette.exceptions import HTTPException as StarletteHTTPException

from app.api import (
    admin,
    analytics,
    auth,
    bug_report,
    debug,
    dependency,
    issue,
    queue,
    rate,
    reading_orders,
    roll,
    session,
    snooze,
    test_helpers,
    thread,
    undo,
)
from app.cache import cache
from app.config import get_app_settings, get_database_settings, get_redis_settings
from app.csrf import CSRF_COOKIE_NAME, CSRF_HEADER_NAME, is_csrf_protected_request
from app.database import get_db
from app.exception_handlers import register_exception_handlers
from app.lifecycle import init_database
from app.middleware import limiter, SecurityHeadersMiddleware
from app.middleware.request_logging import add_request_logging_middleware

logger = logging.getLogger(__name__)


def _default_log_level(environment: str) -> int:
    """Resolve the default root log level for an environment.

    Production defaults to WARNING so structured slow-request and client-error
    warnings reach deployment logs. Set ``LOG_LEVEL`` to raise or lower it.

    Args:
        environment: Current application environment.

    Returns:
        The default root logging level for the environment.
    """
    env_level_map: dict[str, int] = {
        "production": logging.WARNING,
        "staging": logging.WARNING,
        "development": logging.DEBUG,
        "test": logging.WARNING,
    }
    return env_level_map.get(environment, logging.WARNING)


def _resolve_log_level(environment: str) -> int:
    """Resolve the effective root log level, honoring a ``LOG_LEVEL`` override.

    Args:
        environment: Current application environment.

    Returns:
        The root logging level to configure.
    """
    requested_level = os.getenv("LOG_LEVEL")
    if requested_level:
        return logging.getLevelNamesMapping().get(
            requested_level.upper(),
            _default_log_level(environment),
        )
    return _default_log_level(environment)


def _configure_logging(environment: str) -> None:
    """Configure root logging unless a handler is already installed.

    Uvicorn leaves the root logger without handlers, so this path runs in
    production. ``basicConfig`` is a no-op when handlers already exist.

    Args:
        environment: Current application environment.
    """
    if logging.getLogger().hasHandlers():
        return
    logging.basicConfig(level=_resolve_log_level(environment))


# Log database URL at startup (with password redacted)
_db_settings = get_database_settings()
_redacted_url = make_url(_db_settings.database_url).render_as_string(hide_password=True)
logger.info(f"Starting with DATABASE_URL: {_redacted_url}")

def create_app(*, serve_frontend: bool = True) -> FastAPI:
    """Create and configure the FastAPI application.

    Args:
        serve_frontend: Whether to mount frontend static assets and SPA routes.

    Returns:
        Configured FastAPI application instance.
    """
    app_settings = get_app_settings()

    _configure_logging(app_settings.environment)

    app = FastAPI(
        title="Dice-Driven Comic Tracker",
        description="API for tracking comic reading with dice rolls",
        version="0.1.0",
    )

    # Register rate limiter and handler for both normal and test-mode rate limiting.
    app.state.limiter = limiter

    def rate_limit_exceeded_handler(request: Request, exc: Exception) -> Response:
        """Adapt slowapi handler to FastAPI's exception handler type.

        Args:
            request: FastAPI request object.
            exc: Raised exception.

        Returns:
            Response generated by slowapi's rate limit handler.
        """
        return _rate_limit_exceeded_handler(request, cast(RateLimitExceeded, exc))

    app.add_exception_handler(RateLimitExceeded, rate_limit_exceeded_handler)

    app_settings.validate_production_cors()
    cors_origins = app_settings.cors_origins_list

    app.add_middleware(
        CORSMiddleware,
        allow_origins=cors_origins,
        allow_credentials=False,
        allow_methods=["*"],
        allow_headers=["*"],
    )

    app.add_middleware(SecurityHeadersMiddleware)

    @app.middleware("http")
    async def csrf_middleware(request: Request, call_next):
        """Require a matching CSRF header and cookie on mutating API requests."""
        if os.getenv("TEST_ENVIRONMENT") == "true":
            return await call_next(request)
        if not is_csrf_protected_request(request.method, request.url.path):
            return await call_next(request)

        auth_header = request.headers.get("authorization")
        if not auth_header:
            return await call_next(request)

        csrf_cookie = request.cookies.get(CSRF_COOKIE_NAME)
        csrf_header = request.headers.get(CSRF_HEADER_NAME)
        if (
            not csrf_cookie
            or not csrf_header
            or not secrets.compare_digest(csrf_cookie, csrf_header)
        ):
            return JSONResponse(
                status_code=status.HTTP_403_FORBIDDEN,
                content={"detail": "CSRF token missing or invalid"},
            )

        return await call_next(request)

    # Error-only request logging (body redaction + environment-aware sanitization).
    add_request_logging_middleware(app, app_settings.environment)

    # Global, HTTP, and validation exception handlers (environment-aware logging).
    register_exception_handlers(app, app_settings)

    # API route prefix convention:
    # - Legacy resources are served under /api/* (threads, roll, queue,
    #   rate, snooze, undo, auth, admin, analytics, bug-reports, sessions).
    # - Newer resources are served under the versioned /api/v1/* surface
    #   (dependencies, issues, reading-orders).
    # - /api/v1/auth/* and /api/v1/sessions/* are explicit, tested
    #   compatibility aliases while maintained callers migrate to v1.
    # - Non-production tooling routes (debug, test) are also mounted under
    #   bare /api/* but only in non-production/test environments — they are
    #   intentional exceptions to the versioning rule, not client APIs.
    # Add new client resources under /api/v1/*; do not introduce new bare
    # /api/* routes.
    app.include_router(roll.router, prefix="/api/roll", tags=["roll"])
    app.include_router(admin.router, prefix="/api", tags=["admin"])
    app.include_router(analytics.router, prefix="/api", tags=["analytics"])
    app.include_router(bug_report.router, prefix="/api/bug-reports", tags=["bug-reports"])
    app.include_router(auth.router, prefix="/api/auth", tags=["auth"])
    app.include_router(auth.router, prefix="/api/v1/auth", tags=["auth"])
    app.include_router(thread.router, prefix="/api/threads", tags=["threads"])
    if app_settings.environment != "production":
        app.include_router(debug.router, prefix="/api", tags=["debug"])
    app.include_router(issue.router, tags=["issues"])
    app.include_router(rate.router, prefix="/api/rate", tags=["rate"])
    app.include_router(queue.router, prefix="/api/queue", tags=["queue"])
    app.include_router(reading_orders.router, tags=["reading-orders"])
    app.include_router(session.router, prefix="/api/sessions", tags=["session"])
    app.include_router(session.router, prefix="/api/v1/sessions", tags=["session"])
    app.include_router(snooze.router, prefix="/api/snooze", tags=["snooze"])
    app.include_router(undo.router, prefix="/api/undo", tags=["undo"])
    app.include_router(dependency.router, prefix="/api/v1", tags=["dependencies"])
    if os.getenv("TEST_ENVIRONMENT") == "true":
        app.include_router(test_helpers.router, prefix="/api", tags=["test"])

    def _assert_production_frontend_assets() -> None:
        """Ensure required frontend artifacts exist in production.

        Raises:
            RuntimeError: If required built frontend artifacts are missing.
        """
        if app_settings.environment != "production":
            return

        spa_index = Path("static/react/index.html")
        assets_dir = Path("static/react/assets")
        has_js = any(assets_dir.glob("*.js")) if assets_dir.exists() else False
        has_css = any(assets_dir.glob("*.css")) if assets_dir.exists() else False

        if not spa_index.exists() or not assets_dir.exists() or not has_js or not has_css:
            raise RuntimeError(
                "Missing built frontend artifacts in production. "
                "Expected static/react/index.html and static/react/assets with JS/CSS files."
            )

    @app.api_route(
        "/api/{path:path}",
        methods=["GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS", "HEAD"],
        include_in_schema=False,
    )
    async def api_not_found(path: str) -> JSONResponse:
        """Return a JSON 404 for unknown API routes.

        Args:
            path: API path that was not found.

        Returns:
            JSON response with 404 status code.
        """
        return JSONResponse(status_code=status.HTTP_404_NOT_FOUND, content={"detail": "Not Found"})

    class CacheControlledStaticFiles(StarletteStaticFiles):
        """StaticFiles with explicit cache-control headers for hashed assets."""

        async def get_response(self, path: str, scope: Scope) -> StarletteResponse:
            """Get static file response with cache-control headers.

            Args:
                path: The path to the static file.
                scope: ASGI scope containing request information.

            Returns:
                A Starlette response with cache-control headers for hashed assets.
            """
            response = await super().get_response(path, scope)
            if hasattr(response, "headers"):
                response.headers["Cache-Control"] = "public, max-age=31536000, immutable"
            return response

    if serve_frontend:
        if app_settings.environment == "production":
            _assert_production_frontend_assets()
            app.mount("/static", StaticFiles(directory="static"), name="static")
            app.mount(
                "/assets",
                CacheControlledStaticFiles(directory="static/react/assets"),
                name="assets",
            )
        else:
            if Path("static").exists():
                app.mount("/static", StaticFiles(directory="static"), name="static")
            if Path("static/react/assets").exists():
                app.mount("/assets", StaticFiles(directory="static/react/assets"), name="assets")

        @app.get("/vite.svg")
        async def serve_vite_svg():
            """Serve vite favicon.

            Returns:
                FileResponse with vite.svg file.
            """
            from fastapi.responses import FileResponse

            return FileResponse("static/vite.svg", media_type="image/svg+xml")

        def _serve_spa_index_response() -> Response:
            """Serve SPA index file when available, else return fallback HTML.

            Returns:
                FileResponse for the built SPA index, or fallback HTMLResponse in test environments.
            """
            from fastapi.responses import FileResponse, HTMLResponse

            spa_index = Path("static/react/index.html")
            cache_headers = {"Cache-Control": "no-store, no-cache, must-revalidate"}
            if spa_index.exists():
                return FileResponse(str(spa_index), headers=cache_headers)
            if app_settings.environment == "production":
                raise StarletteHTTPException(status_code=503, detail="Frontend assets unavailable")
            fallback_html = "<!doctype html><html><body><div id='root'></div></body></html>"
            return HTMLResponse(fallback_html, headers=cache_headers)

        @app.get("/")
        async def serve_root():
            """Serve React app at root URL.

            Returns:
                FileResponse with React index.html.
            """
            return _serve_spa_index_response()

        @app.get("/react")
        async def serve_react_redirect():
            """Redirect /react to / for consistent routing.

            Returns:
                RedirectResponse to root URL.
            """
            from fastapi.responses import RedirectResponse

            return RedirectResponse("/", status_code=301)

        @app.get("/react/")
        async def serve_react_redirect_slash():
            """Redirect /react/ to / for consistent routing.

            Returns:
                RedirectResponse to root URL.
            """
            from fastapi.responses import RedirectResponse

            return RedirectResponse("/", status_code=301)

    @app.get("/health", response_model=None)
    async def health_check(db: AsyncSession = Depends(get_db)) -> dict | JSONResponse:
        """Health check endpoint that verifies basic application functionality.

        Returns:
            JSON response with health status and database connection state.
        """
        from sqlalchemy import text

        try:
            await db.execute(text("SELECT 1"))
            return {"status": "healthy", "database": "connected"}
        except sqlalchemy_exc.DBAPIError as e:
            logger.error(f"Health check database connection failed: {e}")
            return JSONResponse(
                status_code=status.HTTP_503_SERVICE_UNAVAILABLE,
                content={"status": "unhealthy", "database": "disconnected", "error": str(e)},
            )

    if serve_frontend:

        @app.get("/{full_path:path}")
        async def serve_react_spa(full_path: str):
            """Serve the React SPA for non-API routes.

            The React app owns routing for paths like /rate, /queue, /history, etc.

            Args:
                full_path: The path to serve.

            Returns:
                FileResponse with React index.html.

            Raises:
                StarletteHTTPException: If path is blocked.
            """
            blocked_prefixes = ("api", "static", "assets", "debug")
            blocked_exact = {"health", "openapi.json", "docs", "redoc", "vite.svg"}

            if (
                full_path in blocked_exact
                or full_path in blocked_prefixes
                or any(full_path.startswith(prefix + "/") for prefix in blocked_prefixes)
            ):
                raise StarletteHTTPException(status_code=404, detail="Not Found")

            return _serve_spa_index_response()

    @app.on_event("startup")
    async def startup_event():
        """Initialize database and cache on application startup."""
        await init_database(app_settings.environment)

        redis_settings = get_redis_settings()
        if redis_settings.is_configured:
            if redis_settings.upstash_redis_rest_url and redis_settings.upstash_redis_rest_token:
                await cache.initialize(
                    url=redis_settings.upstash_redis_rest_url,
                    token=redis_settings.upstash_redis_rest_token,
                )
            elif redis_settings.redis_url:
                await cache.initialize(local_url=redis_settings.redis_url)
        else:
            logger.info("Redis cache not configured - caching disabled")

    @app.on_event("shutdown")
    async def shutdown_event():
        """Close cache connection on application shutdown."""
        await cache.close()

    return app


app = create_app()