# frontend/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "BugReportCreate": {
        "description": "Request schema for creating a bug report.",
        "properties": {
          "description": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "diagnostics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BugReportDiagnostics"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "description"
        ],
        "title": "BugReportCreate",
        "type": "object"
      },
...
      "BugReportResponse": {
        "description": "Response after successfully creating a bug report issue.",
        "properties": {
          "issue_url": {
            "title": "Issue Url",
            "type": "string"
          }
        },
        "required": [
          "issue_url"
        ],
        "title": "BugReportResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RateRequest": {
        "description": "Schema for rating request.",
        "properties": {
          "finish_session": {
            "default": false,
            "title": "Finish Session",
            "type": "boolean"
          },
          "issue_number": {
            "anyOf": [
              {
                "maxLength": 50,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Issue number just read, e.g. '5' or 'Annual 1' (triggers migration if provided)",
            "title": "Issue Number"
          },
          "rating": {
            "maximum": 5.0,
            "minimum": 0.5,
            "title": "Rating",
            "type": "number"
          }
        },
        "required": [
          "rating"
        ],
        "title": "RateRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API for tracking comic reading with dice rolls",
    "title": "Dice-Driven Comic Tracker",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/bug-reports/": {
      "post": {
        "description": "Create a bug report issue on GitHub.\n\nArgs:\n    body: Bug report with title, description, and optional diagnostics\n    current_user: The authenticated user\n\nReturns:\n    Created bug report issue URL\n\nRaises:\n    HTTPException: If validation fails or GitHub integration not configured",
        "operationId": "create_bug_report_api_bug_reports__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BugReportCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BugReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Bug Report",
        "tags": [
          "bug-reports",
          "bug-reports"
        ]
      }
    },
...
    "/api/rate/": {
      "post": {
        "description": "Rate current reading and update its thread.\n\nArgs:\n    request: FastAPI request object for rate limiting.\n    rate_data: Rating request data.\n    current_user: Authenticated user making the request.\n    db: SQLAlchemy session.\n\nReturns:\n    Updated thread response.\n\nRaises:\n    HTTPException: If no active session, rating, or thread is valid.",
        "operationId": "rate_thread_api_rate__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rate Thread",
        "tags": [
          "rate"
        ]
      }
    },
    "/api/roll/": {
      "post": {
        "description": "Roll dice to select a thread.\n\nArgs:\n    roll_request: The roll request.\n    request: FastAPI request object for rate limiting.\n    current_user: The authenticated user making the request.\n    db: SQLAlchemy session for database operations.\n\nReturns:\n    RollResponse with selected thread and die result.\n\nRaises:\n    HTTPException: If no active threads available.",
        "operationId": "roll_dice_api_roll__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RollResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Roll Dice",
        "tags": [
          "roll",
          "roll"
        ]
      }
    },
...
    "/api/sessions/": {
      "get": {
        "description": "List sessions with cursor-based pagination.\n\nArgs:\n    current_user: The authenticated user making the request.\n    page_size: Number of sessions to return per page (default 50, max 200).\n    page_token: Token for pagination continuation (started_at timestamp, session_id).\n    db: SQLAlchemy session for database operations.\n\nReturns:\n    SessionHistoryListResponse with paginated sessions and next_page_token if more exist.",
        "operationId": "list_sessions_api_sessions__get",
        "parameters": [
          {
            "description": "Number of sessions to return per page (default 50, max 200)",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Number of sessions to return per page (default 50, max 200)",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "description": "Token for pagination continuation (started_at,session_id)",
            "in": "query",
            "name": "page_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Token for pagination continuation (started_at,session_id)",
              "title": "Page Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionHistoryListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Sessions",
        "tags": [
          "session",
          "sessions"
        ]
      }
    },
...
    "/api/sessions/{session_id}": {
      "get": {
        "description": "Get single session by ID.\n\nArgs:\n    session_id: The session ID to retrieve.\n    current_user: The authenticated user making the request.\n    db: SQLAlchemy session for database operations.\n\nReturns:\n    SessionResponse with session details.\n\nRaises:\n    HTTPException: If session not found.",
        "operationId": "get_session_api_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Session",
        "tags": [
          "session",
          "sessions"
        ]
      }
    },
...
    "/api/snooze/": {
      "post": {
        "description": "Snooze the pending thread, demote it in the queue, and step the die up.\n\nThis endpoint:\n1. Gets the current session (must exist with a pending_thread_id)\n2. Moves the pending thread beyond the widened roll range\n3. Adds the pending_thread_id to snoozed_thread_ids\n4. Steps the die UP (wider pool) using dice ladder logic\n5. Records a \"snooze\" event\n6. Clears pending_thread_id\n7. Returns the updated session\n\nArgs:\n    request: FastAPI request object for rate limiting.\n    current_user: The authenticated user making the request.\n    db: SQLAlchemy session for database operations.\n\nReturns:\n    SessionResponse containing the updated session with snoozed_thread_ids,\n    cleared pending_thread_id, and current die state.\n\nRaises:\n    HTTPException: If no active session exists or no pending thread to snooze.",
        "operationId": "snooze_thread_api_snooze__post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Snooze Thread",
        "tags": [
          "snooze"
        ]
      }
    },
...
    "/api/threads/": {
      "get": {
        "description": "List threads ordered by position with cursor-based pagination.\n\nArgs:\n    request: FastAPI request object for rate limiting.\n    search: Optional case-insensitive title search filter.\n    page_size: Number of threads to return per page (default 50, max 200).\n    page_token: Token for pagination continuation (queue_position,thread_id).\n    current_user: The authenticated user making the request.\n    db: SQLAlchemy session for database operations.\n\nReturns:\n    QueueThreadListResponse with paginated threads and next_page_token if more exist.\n\nRaises:\n    HTTPException: If a retired ``collection_id`` query parameter is present.",
        "operationId": "list_threads_api_threads__get",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "description": "Number of threads to return per page (default 50, max 200)",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Number of threads to return per page (default 50, max 200)",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "description": "Token for pagination continuation (queue_position,thread_id)",
            "in": "query",
            "name": "page_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Token for pagination continuation (queue_position,thread_id)",
              "title": "Page Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueThreadListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Threads",
        "tags": [
          "threads",
          "threads"
        ]
      },
      "post": {
        "description": "Create a new thread.\n\nArgs:\n    request: FastAPI request object for rate limiting.\n    thread_data: Thread creation data.\n    current_user: The authenticated user making the request.\n    db: SQLAlchemy session for database operations.\n\nReturns:\n    ThreadResponse with created thread details.\n\nRaises:\n    RuntimeError: If failed after max retries.",
        "operationId": "create_thread_api_threads__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThreadCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Thread",
        "tags": [
          "threads",
          "threads"
        ]
      }
    },
...
    "/api/v1/sessions/": {
      "get": {
        "description": "List sessions with cursor-based pagination.\n\nArgs:\n    current_user: The authenticated user making the request.\n    page_size: Number of sessions to return per page (default 50, max 200).\n    page_token: Token for pagination continuation (started_at timestamp, session_id).\n    db: SQLAlchemy session for database operations.\n\nReturns:\n    SessionHistoryListResponse with paginated sessions and next_page_token if more exist.",
        "operationId": "list_sessions_api_v1_sessions__get",
        "parameters": [
          {
            "description": "Number of sessions to return per page (default 50, max 200)",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Number of sessions to return per page (default 50, max 200)",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "description": "Token for pagination continuation (started_at,session_id)",
            "in": "query",
            "name": "page_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Token for pagination continuation (started_at,session_id)",
              "title": "Page Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionHistoryListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Sessions",
        "tags": [
          "session",
          "sessions"
        ]
      }
    },
...
  }
}